Move syn crate into context-macro
