--- crates/syn/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "crayfish-syn"
-version.workspace = true
-rust-version.workspace = true
-edition.workspace = true
-license.workspace = true
-authors.workspace = true
-repository.workspace = true
-
-[dependencies]
-syn = { workspace = true, features = ["fold", "visit", "full"] }
-proc-macro2.workspace = true
-quote.workspace = true
